Backend/requirements.txt
astroid==2.2.5
Django==2.2.1
django-rest-framework==0.1.0
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
numpy==1.16.4
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
wrapt==1.11.1
